Split the derive macro into several files and made various simplifications.
